+++ source/Nuke.Common/Tools/Git/GitTasks.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Linq;
+using Nuke.Core;
 using Nuke.Core.Tooling;
 
 namespace Nuke.Common.Tools.Git
@@ -19,6 +20,39 @@ namespace Nuke.Common.Tools.Git
                 processSettings?.EnvironmentVariables,
                 processSettings?.ExecutionTimeout,
                 processSettings?.RedirectOutput ?? false);
+        }
+        
+        public static bool GitIsDetached()
+        {
+            return GitIsDetached(EnvironmentInfo.WorkingDirectory);
+        }
+        
+        public static bool GitIsDetached(string workingDirectory)
+        {
+            var process = StartProcess(
+                            new GitSettings()
+                                    .SetWorkingDirectory(workingDirectory)
+                                    .SetArguments("symbolic-ref --short -q HEAD"),
+                            new ProcessSettings().EnableRedirectOutput())
+                    .AssertWaitForExit();
+
+            return !process.Output.Any();
+        }
+
+        public static bool GitHasUncommitedChanges()
+        {
+            return GitHasUncommitedChanges(EnvironmentInfo.WorkingDirectory);
+        }
+
+        public static bool GitHasUncommitedChanges(string workingDirectory)
+        {
+            var process = StartProcess(
+                    new GitSettings()
+                        .SetWorkingDirectory(workingDirectory)
+                        .SetArguments("status --short"),
+                    new ProcessSettings().EnableRedirectOutput())
+                .AssertZeroExitCode();
+            return process.Output.Any();
         }
     }
 }