+++ source/Nuke.Common/Tools/GitVersion/GitVersionAttribute.cs
@@ -13,24 +13,34 @@ using Nuke.Core.Tooling;
 namespace Nuke.Common.Tools.GitVersion
 {
     /// <inheritdoc/>
+    /// <summary>
+    /// Implements auto-injection for <see cref="GitVersionTasks"/>.
+    /// <para/>
+    /// <inheritdoc/>
+    /// </summary>
     [PublicAPI]
     [UsedImplicitly(ImplicitUseKindFlags.Default)]
     public class GitVersionAttribute : StaticInjectionAttributeBase
     {
+        [Parameter("Allows to bump the major or minor version.")]
+        public static GitVersionBump? Bump { get; }
+
         public static GitVersion Value { get; private set; }
+
+        public bool DisableOnUnix { get; set; }
 
         [CanBeNull]
         public override object GetStaticValue()
         {
-            if (Value != null)
-                return Value;
-
             // TODO: https://github.com/GitTools/GitVersion/issues/1097
-            if (EnvironmentInfo.IsUnix)
+            if (EnvironmentInfo.IsUnix && DisableOnUnix)
             {
-                Logger.Warn($"{nameof(GitVersion)} does not work in UNIX environments.");
+                Logger.Warn($"{nameof(GitVersion)} is disabled on UNIX environment.");
                 return null;
             }
+            
+            if (Value != null)
+                return Value;
 
             var version = GetVersion();
             var tag = GetTag(version);
@@ -53,13 +63,12 @@ namespace Nuke.Common.Tools.GitVersion
         [CanBeNull]
         private static string GetTag(GitVersion version)
         {
-            if (!EnvironmentInfo.ParameterSwitch("major") &&
-                !EnvironmentInfo.ParameterSwitch("minor"))
+            if (!Bump.HasValue || Bump.Value == GitVersionBump.Patch)
                 return null;
 
             ControlFlow.Assert(version.BranchName.Equals("master"), "Version can only be bumped on master branch.");
 
-            return EnvironmentInfo.ParameterSwitch("major")
+            return Bump.Value == GitVersionBump.Major
                 ? $"{version.Major + 1}.0.0"
                 : $"{version.Major}.{version.Minor + 1}.0";
         }