+++ source/Nuke.Core.Tests/BuildServerTest.cs
@@ -56,6 +56,25 @@ namespace Nuke.Core.Tests
             AssertProperty(Jenkins.Instance.NotNull(), property);
         }
 
+        [BuildServerTheory(typeof(Travis))]
+        [MemberData(nameof(Properties), typeof(Travis))]
+        public void TestTravis(PropertyInfo property)
+        {
+            AssertProperty(Travis.Instance.NotNull(), property);
+            Assert.True(Travis.Instance.Ci);
+            Assert.True(Travis.Instance.ContinousIntegration);
+        }
+
+        [BuildServerTheory(typeof(GitLab))]
+        [MemberData(nameof(Properties), typeof(GitLab))]
+        public void TestGitLab(PropertyInfo property)
+        {
+            AssertProperty(GitLab.Instance.NotNull(), property);
+            Assert.True(GitLab.Instance.Ci);
+            Assert.True(GitLab.Instance.GitLabCi);
+            Assert.True(GitLab.Instance.Server);
+        }
+
         public static IEnumerable<object[]> Properties(Type type)
         {
             return type.GetProperties(BindingFlags.Public | BindingFlags.Instance)