+++ source/Nuke.Core/EnvironmentInfo.Internal.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Linq;
 using Nuke.Core.BuildServers;
+using Nuke.Core.Execution;
 using Nuke.Core.Utilities;
 
 namespace Nuke.Core
@@ -51,7 +52,7 @@ namespace Nuke.Core
                 {
                     Logger.Warn(new[]
                                 {
-                                    "The 'Target' parameter is deprecated. " +
+                                    "The 'Target' parameter is deprecated.",
                                     "Starting with the next version, targets need to be specified positional:",
                                     string.Empty,
                                     "  Usage: build <target1[+target2]> [-parameter value]",
@@ -60,7 +61,29 @@ namespace Nuke.Core
                     return parameter;
                 }
 
-                return new[] { "default" };
+                return new[] { BuildExecutor.DefaultTarget };
+            }
+        }
+        
+        internal static string[] SkippedTargets
+        {
+            get
+            {
+                if (ParameterSwitch("nodeps"))
+                {
+                    Logger.Warn(new[]
+                                {
+                                    "The 'NoDeps' switch is deprecated.",
+                                    "Starting with the next version, you can use the 'Skip' parameter to skip all dependencies or only specified ones:",
+                                    string.Empty,
+                                    "  Usage: build -skip",
+                                    "         build -skip Clean+Restore",
+                                    string.Empty
+                                }.JoinNewLine());
+                    return new string[0];
+                }
+
+                return null;
             }
         }
     }