+++ source/Nuke.Core/Execution/InjectionService.cs
@@ -3,8 +3,10 @@
 // https://github.com/nuke-build/nuke/blob/master/LICENSE
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using Nuke.Core.Utilities.Collections;
 
 namespace Nuke.Core.Execution
 {
@@ -12,12 +14,11 @@ namespace Nuke.Core.Execution
     {
         public static void InjectValues(NukeBuild build)
         {
-            var injectionMembers = build.GetType()
-                .GetMembers(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
-                .Where(x => x.GetCustomAttributes<InjectionAttributeBase>().Any()).ToList();
-
             var anyInjected = false;
 
+            var injectionMembers = build.GetInjectionMembers()
+                    .OrderByDescending(x => x.GetCustomAttribute<ParameterAttribute>() != null);
+            
             foreach (var member in injectionMembers)
             {
                 var attributes = member.GetCustomAttributes().OfType<InjectionAttributeBase>().ToList();
@@ -34,7 +35,7 @@ namespace Nuke.Core.Execution
                 var valueType = value.GetType();
                 ControlFlow.Assert(memberType.IsAssignableFrom(valueType),
                     $"Field '{member.Name}' must be of type '{valueType.Name}' to get its valued injected from '{attribute.GetType().Name}'.");
-                PrivateInvoke.SetValue(build, member, value);
+                ReflectionService.SetValue(build, member, value);
 
                 anyInjected = true;
             }