+++ source/Nuke.Core/ParameterAttribute.cs
@@ -11,8 +11,7 @@ namespace Nuke.Core
 {
     /// <inheritdoc/>
     /// <summary>
-    ///     <inheritdoc/><para/>
-    ///     Parameters are resolved case-insensitively in the following order:
+    ///     Injected parameters are resolved case-insensitively in the following order:
     ///     <ul>
     ///         <li>From command-line arguments (e.g., <c>-arg value</c>)</li>
     ///         <li>From environment variables (e.g., <c>Arg=value</c>)</li>
@@ -20,8 +19,9 @@ namespace Nuke.Core
     ///     <para/>
     ///     For value-types, there is a distinction between pure value-types, and their <em>nullable</em>
     ///     counterparts. For instance, <c>int</c> will have its default value <c>0</c> even when it's not
-    ///     supplied via command-line or environment variable, and therefore also can't be used as requirements.
-    ///     Declaring the field as <c>int?</c> however, will enable validation and setting the requirement.
+    ///     supplied via command-line or environment variable.
+    ///     <para/>
+    ///     <inheritdoc/>
     /// </summary>
     /// <example>
     ///     <code>
@@ -51,9 +51,9 @@ namespace Nuke.Core
         [CanBeNull]
         public override object GetValue(string memberName, Type memberType)
         {
-            memberType = Nullable.GetUnderlyingType(memberType) == null
-                         && memberType != typeof(string)
-                         && !memberType.IsArray
+            memberType = Nullable.GetUnderlyingType(memberType) == null &&
+                         memberType != typeof(string) &&
+                         !memberType.IsArray
                 ? typeof(Nullable<>).MakeGenericType(memberType)
                 : memberType;
             return s_parameterService.GetParameter(Name ?? memberName, memberType, (Separator ?? string.Empty).SingleOrDefault());